Refactor INI classes

+++ src/extensions/warheadtype/warheadtypeext.h
@@ -77,7 +77,7 @@ WarheadTypeClassExtension final : public AbstractTypeClassExtension
         /**
          *  This is used to override the size of the combat light flash at the point of impact.
          */
-        float CombatLightSize;
+        double CombatLightSize;
 
         /**
          *  These values are used to shake the screen when the projectile impacts.
